requirements.txt
click==7.1.2
Flask==1.1.2
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
Werkzeug==1.0.1
click
flask
gunicorn
itsdangerous
jinja2
markupsafe
pillow
requests
werkzeug